+++ packages/frontend/src/components/DashboardTiles/DashboardChartTile.tsx
@@ -11,7 +11,6 @@ import {
     DashboardFilterRule,
     Field,
     fieldId,
-    FilterGroup,
     FilterOperator,
     friendlyName,
     getCustomLabelsFromTableConfig,
@@ -27,9 +26,10 @@ import {
 } from '@lightdash/common';
 import React, { FC, useCallback, useEffect, useMemo, useState } from 'react';
 import CopyToClipboard from 'react-copy-to-clipboard';
-import { useHistory, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { v4 as uuidv4 } from 'uuid';
 import useDashboardFiltersForExplore from '../../hooks/dashboard/useDashboardFiltersForExplore';
+import useSavedQueryWithDashboardFilters from '../../hooks/dashboard/useSavedQueryWithDashboardFilters';
 import { EChartSeries } from '../../hooks/echarts/useEcharts';
 import useToaster from '../../hooks/toaster/useToaster';
 import { downloadCsv } from '../../hooks/useDownloadCsv';
@@ -162,6 +162,34 @@ const ValidDashboardChartTile: FC<{
     );
 };
 
+const ValidDashboardChartTileMinimal: FC<{
+    tileUuid: string;
+    data: SavedChart;
+    project: string;
+}> = ({ tileUuid, data, project }) => {
+    const { data: resultData, isLoading } = useSavedChartResults(project, data);
+    const { data: explore } = useExplore(data.tableName);
+
+    return (
+        <VisualizationProvider
+            minimal
+            chartType={data.chartConfig.type}
+            initialChartConfig={data.chartConfig}
+            initialPivotDimensions={data.pivotConfig?.columns}
+            resultsData={resultData}
+            explore={explore}
+            isLoading={isLoading}
+            columnOrder={data.tableConfig.columnOrder}
+        >
+            <LightdashVisualization
+                isDashboard
+                tileUuid={tileUuid}
+                $padding={0}
+            />
+        </VisualizationProvider>
+    );
+};
+
 const InvalidDashboardChartTile: FC = () => (
     <NonIdealState
         title="No chart available"
@@ -170,12 +198,15 @@ const InvalidDashboardChartTile: FC = () => (
     />
 );
 
-type Props = Pick<
-    React.ComponentProps<typeof TileBase>,
-    'tile' | 'onEdit' | 'onDelete' | 'isEditMode'
-> & { tile: IDashboardChartTile };
+interface DashboardChartTileMainProps
+    extends Pick<
+        React.ComponentProps<typeof TileBase>,
+        'tile' | 'onEdit' | 'onDelete' | 'isEditMode'
+    > {
+    tile: IDashboardChartTile;
+}
 
-const DashboardChartTile: FC<Props> = (props) => {
+const DashboardChartTileMain: FC<DashboardChartTileMainProps> = (props) => {
     const { showToastSuccess } = useToaster();
     const { track } = useTracking();
     const {
@@ -231,7 +262,6 @@ const DashboardChartTile: FC<Props> = (props) => {
         pivotReference?: PivotReference;
     }>();
     const [isCSVExportModalOpen, setIsCSVExportModalOpen] = useState(false);
-    useState(false);
 
     const onSeriesContextMenu = useCallback(
         (e: EchartSeriesClickEvent, series: EChartSeries[]) => {
@@ -307,51 +337,18 @@ const DashboardChartTile: FC<Props> = (props) => {
         },
         [explore, savedQuery],
     );
-    // START DASHBOARD FILTER LOGIC
-    // TODO: move this logic out of component
-    let savedQueryWithDashboardFilters: SavedChart | undefined;
+
+    const { data: savedQueryWithDashboardFilters, dashboardFilters } =
+        useSavedQueryWithDashboardFilters(tileUuid, savedChartUuid);
 
     const dashboardFiltersThatApplyToChart = useDashboardFiltersForExplore(
         tileUuid,
         explore,
     );
 
-    if (savedQuery) {
-        const dimensionFilters: FilterGroup = {
-            id: 'yes',
-            and: [
-                ...(savedQuery.metricQuery.filters.dimensions
-                    ? [savedQuery.metricQuery.filters.dimensions]
-                    : []),
-                ...dashboardFiltersThatApplyToChart.dimensions,
-            ],
-        };
-        const metricFilters: FilterGroup = {
-            id: 'no',
-            and: [
-                ...(savedQuery.metricQuery.filters.metrics
-                    ? [savedQuery.metricQuery.filters.metrics]
-                    : []),
-                ...dashboardFiltersThatApplyToChart.metrics,
-            ],
-        };
-        savedQueryWithDashboardFilters = {
-            ...savedQuery,
-            metricQuery: {
-                ...savedQuery.metricQuery,
-                filters: {
-                    dimensions: dimensionFilters,
-                    metrics: metricFilters,
-                },
-            },
-        };
-    }
-    // END DASHBOARD FILTER LOGIC
-
-    const appliedFilterRules = [
-        ...dashboardFiltersThatApplyToChart.dimensions,
-        ...dashboardFiltersThatApplyToChart.metrics,
-    ];
+    const appliedFilterRules = dashboardFilters
+        ? [...dashboardFilters.dimensions, ...dashboardFilters.metrics]
+        : [];
 
     const renderFilterRule = useCallback(
         (filterRule: DashboardFilterRule) => {
@@ -649,6 +646,57 @@ const DashboardChartTile: FC<Props> = (props) => {
             ) : null}
         </>
     );
+};
+
+const DashboardChartTileMinimal: FC<DashboardChartTileMainProps> = (props) => {
+    const {
+        tile: {
+            uuid: tileUuid,
+            properties: { savedChartUuid },
+        },
+    } = props;
+    const { projectUuid } = useParams<{ projectUuid: string }>();
+
+    const { isLoading, data } = useSavedQueryWithDashboardFilters(
+        tileUuid,
+        savedChartUuid,
+    );
+
+    return (
+        <TileBase
+            title={data?.name || ''}
+            clickableTitle={false}
+            titleHref={`/projects/${projectUuid}/saved/${savedChartUuid}/`}
+            description={data?.description}
+            isLoading={isLoading}
+            {...props}
+        >
+            {data ? (
+                <ValidDashboardChartTileMinimal
+                    tileUuid={tileUuid}
+                    data={data}
+                    project={projectUuid}
+                />
+            ) : (
+                <InvalidDashboardChartTile />
+            )}
+        </TileBase>
+    );
+};
+
+interface DashboardChartTileProps extends DashboardChartTileMainProps {
+    minimal?: boolean;
+}
+
+const DashboardChartTile: FC<DashboardChartTileProps> = ({
+    minimal = false,
+    ...rest
+}) => {
+    if (minimal) {
+        return <DashboardChartTileMinimal {...rest} />;
+    } else {
+        return <DashboardChartTileMain {...rest} />;
+    }
 };
 
 export default DashboardChartTile;